requirements.txt
annotated-types==0.7.0
bcrypt==4.1.3
click==8.1.7
colorama==0.4.6
dnspython==2.6.1
email_validator==2.1.2
exceptiongroup==1.2.1
Faker==25.8.0
idna==3.7
iniconfig==2.0.0
logger==1.4
markdown-it-py==3.0.0
mdurl==0.1.2
packaging==24.1
pluggy==1.5.0
pydantic_core==2.18.4
Pygments==2.18.0
pytest==8.2.2
python-dateutil==2.9.0.post0
rich==13.7.1
six==1.16.0
tomli==2.0.1
typing_extensions==4.12.2
